Disable http2 support
HTTP/2 support introduces a dependency on ALPN, which is tightly bound to the version of Java currently running. The latest stable build of Jetty 9.3 does not include the specific module for the OpenJDK version currently in the yum repository, so we'd have to jump through more hoops to get ALPN working correctly.